Add TypeScript code generator
